feat(results): added export to csv to results page
