+++ angular-src/src/app/components/results/results.component.ts
@@ -6,6 +6,7 @@ import { IMovie } from '../../interfaces/movie';
 import { MatDialog } from '@angular/material';
 import { ModalComponent } from './modal.component';
 import { Router } from '@angular/router';
+import { CsvService } from '../../services/csv.service';
 
 @Component({
     selector: 'app-result',
@@ -21,6 +22,10 @@ export class ResultComponent implements OnInit {
   public resultSub: Subscription;
   public rate;
 
+  // csvdata
+  public csvData: any;
+  private _blob: Blob;
+
     // piechart options
     public showLegend = true;
     public view: any[] = [700, 400];
@@ -33,7 +38,7 @@ export class ResultComponent implements OnInit {
     public pieData: any;
 
 
-    constructor(private _search: SearchService, public dialog: MatDialog, private _router: Router) {}
+    constructor(private _search: SearchService, public dialog: MatDialog, private _router: Router, private _csv: CsvService) {}
 
     ngOnInit(): void {
         this._search.resultsSubscription().subscribe(results => {
@@ -41,11 +46,22 @@ export class ResultComponent implements OnInit {
             this.storedResults = new MatTableDataSource<IMovie>(this._rawResults);
             this.ratings = Array(5); // [0,1,2,3,4]
             this.rate = (r) => (this.ratings = r);
+            const fields:string[] = [];
+            for (let key in results.result) {
+                fields.push[key];
+            }
+            this.csvData = this._csv.create(results.result, fields);
+            this._blob = new Blob ([this.csvData], {
+                type: 'text/csv'
+            });
         });
-
 
         this._search.refreshResults();
     }
+  
+  public exportToCsv() {
+    this._csv.download(this._blob);
+  }
 
     public stringCutoff(string: string, maxLength) {
         if (!string) { return string; } else {
