+++ angular-src/src/app/services/csv.service.ts
@@ -17,4 +17,19 @@ export class CsvService {
         return string;
     }
 
+    public download(csv:Blob) {
+        if (window.navigator && window.navigator.msSaveOrOpenBlob) {
+            window.navigator.msSaveOrOpenBlob(csv, 'my.csv');
+        } else {
+            let a = document.createElement('a');
+            a.href = URL.createObjectURL(csv);
+            const now = new Date();
+            const filename = `TNR_revenue-${now.getHours()}${now.getMinutes()}_${now.getMonth() + 1}${now.getDate()}${now.getFullYear()}.csv`;
+            a.download = filename;;
+            document.body.appendChild(a);
+            a.click();
+            document.body.removeChild(a);
+        }
+    }
+
 }
