+++ angular-src/src/styles.scss
@@ -63,3 +63,7 @@
   .justify-center {
       justify-content: center;
   }
+
+  .justify-around {
+      justify-content: space-around;
+  }
